refactor(triple-curlies): rename to no-triple-curlies
