+++ lib/config/recommended.js
@@ -8,7 +8,7 @@ module.exports = {
     'no-html-comments': true,
     'no-nested-interactive': true,
     'self-closing-void-elements': true,
-    'triple-curlies': true,
+    'no-triple-curlies': true,
     'link-rel-noopener': true,
     'no-invalid-interactive': true,
     'img-alt-attributes': true,
