+++ lib/rules/index.js
@@ -13,7 +13,7 @@ module.exports = {
   'img-alt-attributes': require('./lint-img-alt-attributes'),
   'no-inline-styles': require('./lint-no-inline-styles'),
   'link-rel-noopener': require('./lint-link-rel-noopener'),
-  'triple-curlies': require('./lint-triple-curlies'),
+  'no-triple-curlies': require('./lint-no-triple-curlies'),
   'self-closing-void-elements': require('./lint-self-closing-void-elements'),
   'no-nested-interactive': require('./lint-no-nested-interactive'),
   'inline-link-to': require('./lint-inline-link-to'),
