+++ lib/rules/internal/remove-configuration-html-comments.js
@@ -1,7 +1,7 @@
 'use strict';
 
 // this Babel plugin removes configuration comments such as:
-//   <!-- template-lint triple-curlies=false -->
+//   <!-- template-lint no-triple-curlies=false -->
 //   <!-- template-lint enabled=false -->
 
 const AstNodeInfo = require('../../helpers/ast-node-info');
