+++ lib/rules/lint-no-triple-curlies.js
